% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_skeleton.R
\name{project_skeleton}
\alias{project_skeleton}
\title{Create Project Skeleton Structure}
\usage{
project_skeleton(
  project_path,
  init_packages = c("drake", "tidyverse"),
  init_rproj = TRUE,
  init_git = TRUE,
  init_renv = TRUE,
  init_report = NULL
)
}
\arguments{
\item{project_path}{a character string containing the path where the project should be created.}

\item{init_packages}{a character vector containing names of packages that should be written to project's 'packages.R' file.}

\item{init_rproj}{a logical value indicating whether the created project should also be an Rstudio-style project. Default is \code{TRUE}.}

\item{init_git}{a logical value indicating whether the created project should also be tracked by git. Default is \code{TRUE}.}

\item{init_renv}{a logical value indicating whether the created project should use \code{renv} for package management. Default is \code{TRUE}.}

\item{init_report}{a character string indicating which of the report templates available in \code{hsRd} should be included in the project folder 'reports'.
If \code{NULL} (the default), then no template is added.}
}
\value{
Returns \code{NULL} silently.
}
\description{
\code{project_skeleton} creates a project structure that's particularly useful for drake projects.
}
\details{
The folder path specified in \code{project_path} should be a path of directories ending in the name of the not-yet-existing folder that will contain
  the project. For instance, to create a project named \emph{my_analysis} inside the folder \emph{my_project} on the \emph{J} drive, you would specify
  \code{project_path = "J:/my_project/my_analysis"} in the call to \code{project_skeleton}. This function creates folders and files, so you must have write
  permissions within the directory that contains your project.
}
